move inner classes to dedicated package
